+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Adds support for the Role endpoint for searching and paging Role Members
 
+* Adds additional escaping to URL parameters on requests to handle special characters (e.g. spaces)
+
 # v5.1.0
 
 * Introduces backwards compatibility with Conjur 4.x for most API methods.